+++ src/app/services/websocket-supplier/websocket-supplier.service.ts
@@ -1,4 +1,4 @@
-export class WebSocketSupplier {
+export class WebSocketSupplierService {
   private _socket = io.connect('http://localhost:3000');
 
   get(): any {
